+++ changelog.md
@@ -2,5 +2,6 @@ Changelog
 
 ## v1.22 (UNRELEASED)
 
+  + Added more options 'topic', 'qos' and 'retain' to MQTT returner.
   + Possible to enable Bluetooth on the hardware UART (RPi default) for devices not using STN chip.
   + Support sync of grains defined in pillar data. This can be used to override local QMI settings from the cloud.